templates/requirements.txt
Flask==2.3.4
python-dotenv==1.0.0
mysql-connector-python==8.1.0
requests==2.31.0
flask-cors==4.0.1
